+++ source/_static/js/redirects.js
@@ -75,6 +75,8 @@ newUrls['4.12'] = [
   '/development/packaging/generate-server-package.html',
   '/development/packaging/generate-agent-package.html',
   '/cloud-service/ai-analyst.html',
+  '/deployment-options/docker/prerequisites.html',
+  '/deployment-options/docker/build-docker-images-locally.html',
 ]
 
 /* Redirections from 4.9 to 4.10  */
@@ -125,6 +127,11 @@ redirections.push(
     '4.11': '/user-manual/capabilities/vulnerability-detection/offline-update.html',
     '4.12': '/user-manual/capabilities/vulnerability-detection/configuring-scans.html#offline-vulnerability-detection',
   },
+  {
+    'target': ['4.11=>4.12', '4.12=>4.11'],
+    '4.11': '/deployment-options/docker/wazuh-container.html#build-docker-images-locally',
+    '4.12': '/deployment-options/docker/build-docker-images-locally.html',
+  },
 );
 
   /* Pages no longer available in 4.12 */
@@ -138,6 +145,8 @@ removedUrls['4.12'] = [
   '/development/packaging/generate-hpux-package.html',
   '/development/packaging/generate-wpk-package.html',
   '/user-manual/capabilities/vulnerability-detection/offline-update.html',
+  '/deployment-options/docker/docker-installation.html',
+  '/deployment-options/docker/faq-wazuh-container.html',
 ];
 
 /* *** RELEASE 4.11 ****/